Don't append '_sub' to checks by default
This seems to have mostly been added to avoid the nagios nrpe charm
stomping on checks shipped out by the nagios charm (on the same
unit). This changes the default to not append _sub to checks and
detect if this nrpe subordinate charm is also running on a a nagios
unit.