Allow custom local checks to be specified
Some user environments have local checks which need to be
configured for Nagios monitoring but are not available in
the nrpe charm. Some of these checks may very dependent on
the environment they are deployed in.

This change adds the ability to define custom local checks
to the nrpe-charm configuration.

Co-Authored-By: Matt Rae <matt.rae@canonical.com>
Closes-Bug: #1687116