allow direct includes from /etc/rsync-juju.d
This allows *.inc files to be added that override global settings.

For example, the ubuntu-mirror charm wants to be able to set "motd file",
and currently it can only do this by overwriting /etc/rsyncd.conf.

The *.conf files processed by &include can set defaults at top-level
that affect the rest of the modules in that file, but they cannot
affect the rest of the configuration, or change global options.