Add python-openstackclient to functests
Since b46f13ef113e690493364dabc87821699626d8e6 in zaza, if running tests
on an OpenStack cloud provider, each bundle test concludes with a
cleanup routine that needs the python-openstackclient package in
requirements.txt for the tox environment, plus the OS_ environment vars
exported/passed through such that the Openstack client actually works.
If these are not set, the test fails on the first bundle cleanup and
does not continue to the remaining bundles.