Reworked some checks to only run on hosts
Some checks have limited to no meaning when run on containers.  This
patch tries to change some of those checks so they're only enabled and
installed on host machines.

Do note that this does impact testing.  Functional tests expect that
the main machines of the model are not containers, and a newly added
test for this functionality will fail if it is run on an LXD-based
Juju controller.