Bind NRPE to local address
In cases where we specify to use public address and/or Juju returns a
public address regardless of the setting, we cannot use that address for
binding the NRPE service.

This change ensures get_local_ingress_address returns the local address
for binding the nrpe service, and if configured to do so returns the
public address for relation data and NRPE host exports.