Add target-address to relation data
The Nagios charm makes use of a relation-data item named target-address,
which is used for connections from Nagios to the remote unit (NRPE).
If target-address is not set, Nagios then uses ingress-address, and if
that's not set, private-address.  Both ingress-address and
private-address are automatically set by Juju, but updated by the nrpe
charm.

This change adds target-address to the relation data which is controlled
by the NRPE charm from the start, as under some circumstances the
ingress-address is overwritten by Juju.