Use rsync only if required
Only install rsync if export_nagios_definitions=True.
Do not use rsync if nagios_master is set but
export_nagios_definitions=False.

Previously if nagios_master is set, the charm behaves as if
export_nagios_definitions=True, however there are circumstances
requiring nagios_master to be set that do not require the rsync service
to be configured.

Fixes-bug: LP:#1964027