MemAvailable should not be added to Cache memory.
MemAvailable includes MemFree, SReclaimable,
the size of the file LRU lists, and now gets added
to the total size of Cache memory. This can result
in a cache value that is way too large, and in the
rest of the calculation can then yield a negative
used-memory result, and/or a free-memory result that
is higher than the total amount of physical memory.