nrpe_helpers: set default cpu governor checks
From now on we are setting a default check of the value "performance"
for selected principle charms within the nrpe charm code. The old
behavior was to get this default value through the
nrpe-external-master relation.

One side effect to note is that `juju config` will report empty
configuration, however that was also the case for the previous
implementation. This new approach is saving us from updating every
single charm code that we want to introduce a default behavior.

Signed-off-by: Mert Kırpıcı <mert.kirpici@canonical.com>