Fix cis-audit scripts for focal and jammy (#174)
* Fix cis-audit scripts for focal and jammy

Starting with focal, the cis-audit was replaced with usg audit.
Additionally, there were some minor changes in paths and names. This
commit adds functionality to work on focal and jammy.
Closes-Bug: #2041517

* Added function to get the major version from /etc/os-release

Previously, the cron and checks just looked for the existence of the
cis-audit binary. This commit changes this to read the major version
from /etc/os-release.

* imporved unittests for cis audit checks

This commit changes the unittests to be aware of the tooling differences
in cis-auditing between bionic and focal.

* Improved testing and the _get_major_version function

* Exception if VERSION_ID not found and remove readlines()

* Fixing black version in tox.ini

* typo

* Fix lint errors on latest black version

* Remove all version pins on black

---------

Co-authored-by: Marcus Boden <marcus.boden@canonical.com>