Add tailoring file for cis audit
Added the new config cis_audit_tailoringfile to pass a custom
tailoring file for CIS audit.

This new option is mutual exclusive with cis_audit_profile,
meaning that you cannot use both at the same time.

When the tailoring file content is passed, the charm will write
the content at the default place on /etc/usg/default-tailoring.xml

If the option is not used, the charm will ensure that this file
doesn't exist to not interfere with the profile passed.

See the docs for more details:
[0] https://ubuntu.com/security/certifications/docs/usg/cis/customization
[1] https://ubuntu.com/blog/cis-security-compliance-usg