Make status consistent (#223)
Merge the status setting code from `manage()` and `update_status()`,
so the status output is consistent across all hooks.

Also remove the `get_revision()` function, which appears to
have not been working since 734d7f4, when the version file was dropped.
It's possible to link the deployed charm to the source code revision,
by matching the charm revision with the `revNNN` tags in the git source
(added automatically by the release workflow in github).

Also remove old files and update tox to fix tests.

Fixes: #217